Create ReservedModelStates
