+++ modelstore/storage/local.py
@@ -22,8 +22,8 @@ from modelstore.storage.blob_storage import BlobStorage
 from modelstore.storage.util.paths import (
     MODELSTORE_ROOT_PREFIX,
     get_model_state_path,
-    is_valid_state_name,
 )
+from modelstore.storage.states.model_states import is_valid_state_name
 from modelstore.storage.util.versions import sorted_by_created
 from modelstore.utils.log import logger
