+++ modelstore/storage/util/paths.py
@@ -97,16 +97,3 @@ def get_model_state_path(root_dir: str, state_name: str) -> str:
     """
     model_states = get_model_states_path(root_dir)
     return os.path.join(model_states, f"{state_name}.json")
-
-
-def is_valid_state_name(state_name: str) -> bool:
-    if any(state_name == x for x in [None, ""]):
-        logger.debug("state_name has invalid value: %s", state_name)
-        return False
-    if len(state_name) < 3:
-        logger.debug("state_name is too short: %s", state_name)
-        return False
-    if os.path.split(state_name)[1] != state_name:
-        logger.debug("state_name cannot be a path: %s", state_name)
-        return False
-    return True
